+++ examples/next/pages/ui/components/storage/file-uploader/process-file-access-level/index.page.tsx
@@ -9,7 +9,13 @@ Amplify.configure(awsExports);
 const processFile: FileUploaderProps['processFile'] = async ({ file }) => {
   const fileExtension = file.name.split('.').pop();
 
-  return file
+  // pretend the input `file` has been compressed:
+  const blob = new Blob(['Compressed data'], { type: 'text/plain' });
+  const compressedFile = new File([blob], undefined, {
+    type: 'text/plain',
+  });
+
+  return compressedFile
     .arrayBuffer()
     .then((filebuffer) => window.crypto.subtle.digest('SHA-1', filebuffer))
     .then((hashBuffer) => {
@@ -17,7 +23,10 @@ const processFile: FileUploaderProps['processFile'] = async ({ file }) => {
       const hashHex = hashArray
         .map((a) => a.toString(16).padStart(2, '0'))
         .join('');
-      return { file, key: `${hashHex}.${fileExtension}` };
+      return {
+        file: compressedFile,
+        key: `${hashHex}.${fileExtension}`,
+      };
     });
 };